requirements.txt
langchain
langchain-openai
fastapi
uvicorn
python-dotenv
sse-starlette
langserve
